Move ConversionUtil to converter package.
