+++ annotation-rest/src/test/java/uk/ac/ebi/quickgo/annotation/converter/AnnotationToGAFTest.java
@@ -1,9 +1,7 @@
 package uk.ac.ebi.quickgo.annotation.converter;
 
-import uk.ac.ebi.quickgo.annotation.converter.AnnotationToGAF;
 import uk.ac.ebi.quickgo.annotation.model.Annotation;
 import uk.ac.ebi.quickgo.annotation.model.AnnotationMocker;
-import uk.ac.ebi.quickgo.annotation.model.ConversionUtil;
 
 import org.junit.Before;
 import org.junit.Ignore;
