+++ annotation-rest/src/test/java/uk/ac/ebi/quickgo/annotation/converter/AnnotationToGPADTest.java
@@ -2,7 +2,6 @@ package uk.ac.ebi.quickgo.annotation.converter;
 
 import uk.ac.ebi.quickgo.annotation.model.Annotation;
 import uk.ac.ebi.quickgo.annotation.model.AnnotationMocker;
-import uk.ac.ebi.quickgo.annotation.model.ConversionUtil;
 
 import org.junit.Before;
 import org.junit.Test;
